requirements.txt
PyGithub>=2.1.1
PyYAML>=6.0.1
python-dotenv>=1.0.0 